Bank deposit program
The user in this program can find out how much money he will receive if he makes a bank deposit. The user is asked how much he invests for how many years at what percentage and with what amount of replenishment for subsequent years.The function with the replenishment amount does not work correctly, but the error will be fixed soon! Developer:  anasteyshen23